Follow best practice for theme setup
I was following a mixture of old and new practices for theme setup. The
lastest readme for the theme I am using (broccolini/swiss) gives the
correct way of doing it using the `gh-pages` gem, which is what I have
done here.

I'm hoping this will also solve the problem where links are not
underlined on the article pages. This was updated in a newish version of
the theme gem.