Make the post a little more like Ben's
I remembered that I _really_ liked the cumulative count that Ben was
making across all the links (e.g.
http://noisydecentgraphics.typepad.com/design/2018/05/design-research-urls-14.html)
so I updated this to do the same.

I had to make them `h4` to make it look reasonably decent, not quite as
stark as how he has done it but that's the theme I guess.