Another post with links
Getting into the swing of this. I've found a reasonably good method of
doing it.

I also changed the style of the links, making the heading a link for
each one. That works particularly well with this theme as a link has a
dotted line under it which clearly demarcates it from a heading.

Haven't decided whether to go back and change it for #1.