Enable google analytics
To enable google analytics on the site I have temporarily forked the
swiss theme to `kalbir/swiss` and added the code to make GA work there.
This config is what is then needed to pass the google analytics token to
the theme.

I'm hoping to get the google analytics code added to the the main theme
through a PR.