Add explicit excerpting
Before it looks like I was using jekyll to work out where to excerpt a
post on the front page. By adding a setting in config I can go through
each page and add an explicit point at which to excerpt (which I have
made the beginning of each paragraph).