Create a new theme
I probably should have done this in smaller segments but hey I was on a
roll.

What I have done here is:

1. Use the tachyons framework. So I've put the css for that in
`style.scss` and I also pull it from the internet. Coz I'm an ops
person.

2. Written a new homepage. It's very simple: a header, a bit of header
text, and then all of the posts. No pagination, no menus, no links, no
footer. This is based around examples from the tachyons component
library. The base for the page is in `_layouts/default.html` and the
main bit of the page is `index.html`.

3. Written templates for posts and pages. These are again based around
quite a nice looking tachyons example.

4. Added new content to the "about" page.

All of that results in a very very simple and clean, more or less single
page, blog style layout.