Make RSS support explicit
I've been using `jekyll-feed` to generate an atom feed but it hasn't
been very explicit a) how to get the feed and b) that a feed exists (for
crawlers).

To fix these two things I've:

1. Added a link to the feed on the homepage. Maybe this is a bad idea as
it is quite precious real estate but it looks okay at the moment.

2. Added an html header for the rss feed. This is auto generated by the
`jekyll-feed` plugin.

These both seem like good moves but I _might_ revert 1.