Change the default fonts
I've been wanting to djuzz things up with some better fonts for a while.
I think I'm going to end up buying some nice onces but in the meantime
I'm using two google fonts to get the kind of effect I'm looking for.

These are:

- Lora. A serif display font with a little old world character. This is
nice for the headings. I _particularly_ like it for the main h1 and the
titles of the individual article pages. I'd maybe consider using a
second typeface for the <h1 headings.

- Karla. A serif with a bit of air to it and soft, almost squishy
curvature. I really like the "S" and "2" on this one.

An initial improvement.